refactor: moved signup_view to views.py
