+++ multi_tenancy/views.py
@@ -1,11 +1,57 @@
 from posthog.urls import render_template
 from django.http import JsonResponse, HttpResponse
 from django.conf import settings
+from django.contrib.auth import login
+from posthog.models import User, Team
 from django.shortcuts import redirect
 from posthog.api.user import user
 from multi_tenancy.models import TeamBilling
 from multi_tenancy.stripe import create_subscription, customer_portal_url
 import json
+import posthoganalytics
+
+
+def signup_view(request):
+    if request.method == "GET":
+        if request.user.is_authenticated:
+            return redirect("/")
+        return render_template("signup.html", request)
+    if request.method == "POST":
+        email = request.POST["email"]
+        password = request.POST["password"]
+        company_name = request.POST.get("company_name")
+        is_first_user = not User.objects.exists()
+        try:
+            user = User.objects.create_user(
+                email=email, password=password, first_name=request.POST.get("name")
+            )
+        except:
+            return render_template(
+                "signup.html",
+                request=request,
+                context={
+                    "error": True,
+                    "email": request.POST["email"],
+                    "company_name": request.POST.get("company_name"),
+                    "name": request.POST.get("name"),
+                },
+            )
+        team = Team.objects.create_with_data(users=[user], name=company_name)
+        login(request, user, backend="django.contrib.auth.backends.ModelBackend")
+        posthoganalytics.capture(
+            user.distinct_id,
+            "user signed up",
+            properties={"is_first_user": is_first_user},
+        )
+        posthoganalytics.identify(
+            user.distinct_id,
+            properties={
+                "email": user.email,
+                "company_name": company_name,
+                "name": user.first_name,
+            },
+        )
+        return redirect("/")
 
 
 def user_with_billing(request):
